Increase maximum number of stall intervals
This increases the number of "stalls" that occur (the last count bytes
downloaded doesn't change) before we decide that the connection has
stalled.

When connecting to the NASA (SSL-ified) NTrip server the stall detection
was triggered, but it really just need a little more time for the
connection to setup.