Refactor leds_set and brightness_cache for flush
Replaces cache with fully formed state structs so
the cache itself can be passed to leds_set in order
to 'flush' the values out to the driver. leds_set
has retry looping with logging on error or timeout.
Flush logic is to set flush flag if any single write
is not successful, and flush occurs before any call
to update leds when flag is set, then resets the
flag.

swift-nav/piksi_v3_bug_tracking#997

DEVC-42 #done merged in v1.4.0-release and about to merge master