Refactor zmq_adapter
Replaced zsock usage with endpoints, but added
a zsock_read style method to mirror the usage as
it previously existed for drop in replacement.

Refactored req/rep loop to use a pk_loop instead
of a manual loop with zpoll_items. Timeout timers
are created dynamically with each request received
on the reply endpoint and passed between callbacks
using a custom context.