Fork build into INTERNAL, PROTECTED and RELEASE versions
This change adds serveral new make build targets:

- image-release-open: Locks down the image (no SSH, or console access
via USB serial), but the image is NOT encrypted.  This target generates
a filename that follows existing conventions since this is intended to
be the generally available release product for the 1.6 release.  It will
be named per the usual form: `PiksiMulti-v<XYZ>.bin`.

- image-release-protected: Locks down the image (no SSH, or console
access via USB serial), but the image *IS* encrypted.  SmoothPose INS is
not pulled into the image.

- image-release-ins: Locks down the image (no SSH, or console access via
USB serial), but the image *IS* encrypted.  SmoothPose INS *IS* pulled
into the image, access to the SmoothPose git repo (which houses
pre-built binaries) is required in order to build the images.  The image
will be tagged with `PROTECTED`, that is, the file name will take the
following form `PiksiMulti-PROTECTED-v<XYZ>.bin`

The `image` target generates largely the same image, but it will be
tagged with `INTERNAL` instead of the usual naming scheme.  That is:
`PiksiMulti-INTERNVAL-v<XYZ>.bin`

We also introduce several packages that accomplish different things:

- piksi_dev_tools: Moves `dropbear`, `blackmagic`, `strace`, and
`stress` into their own packages so that they can be selectively
enabled (some of these are GPLv3, and cannot be released publicly to
remain compliant).

- release_lockdown: Kind of the inverse of the `piksi_dev_tools`
package, this package drops in several config files that disable USB
UARTs and disables the getty process that services them.