Minimize sends
Make endpoint_adapter minimize sends by aggregating data if it can
during a fixed period of time.  This allows endpoint_router to minimize
the number of kernel transitions it needs to make.

Add a fixed number of packets that we'll service in endpoint_router,
this prevents us from starving other packets on the system if one packet
is particularly active (since nanomsg is single threaded).